Move vg sim into a subcommand
